requirements.txt
accelerate==0.25.0
diffusers==0.27.2
einops==0.7.0
gradio==4.26.0
gradio_client==0.15.1
librosa==0.10.1
numpy==1.26.4
omegaconf==2.3.0
pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
soundfile==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchlibrosa==0.1.0
torchvision==0.15.2
tqdm==4.66.2
